nitroplast_pipeline/requirements.txt
# Core deep learning
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0

# Protein handling
biopython>=1.81
fair-esm>=2.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Clustering for data splitting
cd-hit  # Note: Install via conda or system package manager
mmseqs2  # Note: Install via conda or system package manager

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
umap-learn>=0.5.3

# Progress and logging
tqdm>=4.65.0
wandb>=0.15.0  # Optional: for experiment tracking
tensorboard>=2.13.0

# Configuration
PyYAML>=6.0
hydra-core>=1.3.0

# Utilities
joblib>=1.3.0
scipy>=1.11.0
